requirements.txt
imgviz==1.5.0
keras==2.9.0
kornia==0.6.4
matplotlib==3.5.1
monai==0.8.1
numpy>=1.22
pandas==1.4.1
pathos==0.2.9
Pillow==9.1.1
PyYAML==6.0
scipy==1.7.3
seaborn==0.11.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
wandb==0.12.17
